List optimization
Removed bottom border from cards.  They now alternate between a light
and dark background.

Changed the structure of cards themselves.  It's an improvement, but
still not exactly what I want.

Bug fix:  A random " no longer appears in the topbar / player area.
This was caused by an extra " that got thrown in somewhere.  I blame
cats.  Not the animal,  but the broadway play.  It's just too good.